refactor(api-gateway): file renamed used '-' dash sepration intead of '_' underscore; created index.js file to manage consolidated export
