+++ API-Gateway/src/middlewares/index.js
@@ -1,0 +1,2 @@
+export * from './auth';
+export * from './password';
