+++ API-Gateway/src/middlewares/password.js
