+++ API-Gateway/src/services/ft-commitment.js
