+++ API-Gateway/src/services/nft-commitment.js
